+++ playlist_bridge/__init__.py
@@ -1,5 +1,5 @@
 """Playlist Bridge package."""
 
-__version__ = "2.1.0-beta.5"
+__version__ = "2.1.0-beta.6"
 
-__build__ = "20260917.5"
+__build__ = "20260917.6"